+++ lib/x509/date_time.ex
@@ -23,6 +23,32 @@ defmodule X509.DateTime do
     {:generalizedTime, '#{date}#{time}Z'}
   end
 
+  def to_datetime({:utcTime, time}) do
+    "20#{time}" |> to_datetime()
+  end
+
+  def to_datetime({:generalizedTime, time}) do
+    time |> to_string() |> to_datetime()
+  end
+
+  def to_datetime(
+        <<year::binary-size(4), month::binary-size(2), day::binary-size(2), hour::binary-size(2),
+          minute::binary-size(2), second::binary-size(2), "Z"::binary>>
+      ) do
+    %DateTime{
+      year: String.to_integer(year),
+      month: String.to_integer(month),
+      day: String.to_integer(day),
+      hour: String.to_integer(hour),
+      minute: String.to_integer(minute),
+      second: String.to_integer(second),
+      time_zone: "Etc/UTC",
+      zone_abbr: "UTC",
+      utc_offset: 0,
+      std_offset: 0
+    }
+  end
+
   # Shifts a DateTime value by a number of seconds (positive or negative)
   defp shift(%DateTime{} = datetime, seconds) do
     datetime