+++ test/x509/crl/entry_test.exs
@@ -1,0 +1,4 @@
+defmodule X509.CRL.EntryTest do
+  use ExUnit.Case
+  doctest X509.CRL.Entry
+end