+++ test/x509/crl/extension_test.exs
@@ -1,0 +1,4 @@
+defmodule X509.CRL.ExtensionTest do
+  use ExUnit.Case
+  doctest X509.CRL.Extension
+end